Fix react-refresh warning by moving ThemeProviderWrapper to its own file
ThemeProviderWrapper was defined but not exported from main.tsx,
which react-refresh/only-export-components flags since Fast Refresh
can't work in a file with no exports. Moved it to its own module and
rebuilt player/dist.
